refactored lib.rs. Mod nbt_tag added
