+++ Cargo.toml
@@ -1,12 +1,12 @@
 [package]
 name = "rnbt"
 version = "0.1.0"
-authors = ["caelunshun <caelunshun@gmail.com>"]
-edition = "2018"
+authors = ["caelunshun <caelunshun@gmail.com>", "mrmarkolinus"]
+edition = "2021"
 
 [dependencies]
 byteorder = "1.3"
-derive-new = "0.5"
+derive-new = "0.6.0"
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
 pyo3 = { version = "0.20", features = ["extension-module"] }
